feat: Improve auth (#28)
- add required auth to all necessary methods
- implement mocks for provider

Since some Keystone endpoints have different auth requirements it cannot
be handled in the middleware and should be instead processed with
extractors. Access to the state from extractor is not supporting state
with generics. Due to that it is required to again rework the state, but
now we finally have the proper mock support.