feat: Add token issued_at property (#354)
`issued_at` property of the token in python Keystone is a timestamp of
the fernet creation (part of the protocol). Due to the absence of the
method in the fernet crate copy the logic (just a few steps) how to
access it from the message (python keystone does it the same way even
the public method exists).