refactor: Start passing `ServiceState` into providers (#366)
In order to make providers easier it makes much more sense to pass the
overall app state instead of database connection and provider manager.
This way every backend is able to extract only the information it really
needs and we do not need to extend methods once we figure out that we
need to pass something else.